選手一覧画面のvueコンポーネントの骨組みを作成
