+++ app/javascript/all_teams.js
@@ -6,15 +6,15 @@
 // All it does is render <div>Hello Vue</div> at the bottom of the page.
 
 import Vue from 'vue'
-import App from '../app.vue'
+import AllTeams from './all_teams.vue'
 
 document.addEventListener('DOMContentLoaded', () => {
-  const app = new Vue({
-    render: h => h(App)
-  }).$mount()
-  document.body.appendChild(app.$el)
-
-  console.log(app)
+  const allTeams = document.getElementById('js-all-teams')
+  if (allTeams) {
+    new Vue({
+      render: h => h(AllTeams)
+    }).$mount('#js-all-teams')
+  }
 })
