--- app/javascript/app.vue
@@ -1,22 +1,0 @@
-<template>
-  <div id="app">
-    <p>{{ message }}</p>
-  </div>
-</template>
-
-<script>
-export default {
-  data: function () {
-    return {
-      message: "Hello Vue!"
-    }
-  }
-}
-</script>
-
-<style scoped>
-p {
-  font-size: 2em;
-  text-align: center;
-}
-</style>
